Sat, Aug 10, 2024, 3:34 PM -03:00

+++ credit/credit.c
@@ -5,7 +5,7 @@ int main() {
     long int cardNumber = get_long("Number: ");
     int cksum1 = 0;
     int cksum2 = 0;
-    int multiplier = 1;
+    long int multiplier = 1;
     printf("%li\n", cardNumber);
     for(int i = 0; i < 8; i++) {
